% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{realGARCH_select}
\alias{realGARCH_select}
\title{realGARCH_select}
\usage{
realGARCH_select(gp, insample = F)
}
\arguments{
\item{gp}{is high frequency data}
}
\value{
text
}
\description{
select distribution of innovation along with the realizedVol in realGARCH based on specified methods
}
\details{
#' This is some descriptio of this function.
}
